updated_homelab_topology_internal
The discord bot will be hosted on our VPS and it will use ICMP pings to our homelab’s IP address to carry out a TCP handshake and ensure the homelab is up. If the discord bot fails to carry out the TCP handshake and returns to sender with errors, the bot will ping the users discord alerting its owner that the service is down.
It will continue to ping 2 more times at intervals of 10 minutes. After these 3 consecutive failed pings it will go back to its normal 30 interval timer for pinging the server site unless the server is still down of course.